fix: Рендерит Сортировку и Фильтры отдельно. Выносит рендеринг шапки в отдельный презентер
